+++ packages/rx_bloc_cli/bin/compile_bundles.sh
@@ -36,6 +36,11 @@ dart run mason_cli:mason bundle \
 dart run mason_cli:mason bundle \
   -t dart \
   -o lib/src/templates/ \
+  mason_templates/bricks/lib_translations
+
+dart run mason_cli:mason bundle \
+  -t dart \
+  -o lib/src/templates/ \
   mason_templates/bricks/lib_pin_code
 
 dart run mason_cli:mason bundle \