+++ packages/rx_bloc_cli/lib/src/models/configurations/feature_configuration.dart
@@ -3,6 +3,7 @@ class FeatureConfiguration {
   /// Feature Configuration constructor
   FeatureConfiguration({
     required this.changeLanguageEnabled,
+    required this.remoteTranslationsEnabled,
     required this.analyticsEnabled,
     required this.pushNotificationsEnabled,
     required this.realtimeCommunicationEnabled,
@@ -25,6 +26,9 @@ class FeatureConfiguration {
 
   /// Change language
   final bool changeLanguageEnabled;
+
+  /// Remote translations
+  final bool remoteTranslationsEnabled;
 
   /// Dev menu
   final bool devMenuEnabled;