+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/feature_splash/services/splash_service.dart
@@ -1,17 +1,17 @@
 {{> licence.dart }}
 
-import '../../lib_permissions/services/permissions_service.dart';
-import '../../lib_translations/services/translations_service.dart';
+import '../../lib_permissions/services/permissions_service.dart';{{#enable_remote_translations}}
+import '../../lib_translations/services/translations_service.dart';{{/enable_remote_translations}}
 
 class SplashService {
   SplashService(
-    PermissionsService permissionsService,
-    TranslationsService translationsService,
-  )   : _permissionsService = permissionsService,
-        _translationsService = translationsService;
+    PermissionsService permissionsService,{{#enable_remote_translations}}
+    TranslationsService translationsService,{{/enable_remote_translations}}
+  )   : _permissionsService = permissionsService{{#enable_remote_translations}},
+        _translationsService = translationsService{{/enable_remote_translations}};
 
-  final PermissionsService _permissionsService;
-  final TranslationsService _translationsService;
+  final PermissionsService _permissionsService;{{#enable_remote_translations}}
+  final TranslationsService _translationsService;{{/enable_remote_translations}}
 
   bool _appInitialized = false;
 
@@ -22,8 +22,8 @@ class SplashService {
   }
 
   List<Future<void>> get _nomenclatures => [
-        _permissionsService.load(),
-        _translationsService.load(),
+        _permissionsService.load(),{{#enable_remote_translations}}
+        _translationsService.load(),{{/enable_remote_translations}}
       ];
 
   bool get isAppInitialized => _appInitialized;