+++ packages/rx_bloc_cli/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [5.3.0]
 * Updated project dependencies
+* Added `--enable-remote-translations` flag to configure remote translation lookup for localizations
 
 ## [5.2.0]
 * Update generated project min supported iOS version to `16.0`